Minor modifs + moved things around
